test(import): port chunked-import real-store tests to embedded Dolt
cmd/bd/import_chunking_test.go imported internal/storage/sqlite, which is not
present on main, so `go test -c ./cmd/bd` failed to build (PR Risk: "Build
embedded cmd test binary") and blocked the entire cmd/bd test binary. The same
file also tripped the testing.Short boundaries policy check (PR policy) via
TestImportChunkedConcurrentAvailability.

Split the suite along the store it needs:

- import_chunking_test.go keeps the mock/pure tests (chunkRecordingStore,
  ordering, transaction-boundary counting) and drops the sqlite/filepath/sync
  imports so it always compiles.
- import_chunking_embedded_test.go (//go:build cgo, gated on
  BEADS_TEST_EMBEDDED_DOLT) ports the real-store tests to embeddeddolt.Open --
  the in-process store the repo actually ships -- following the existing
  cmd/bd/*_embedded_test.go convention. Every chunking invariant (durable
  committed prefix after a crash, no ready-window at any freeze point, waits-for
  gating, concurrent-update edge survival, stale-rejection) passes unchanged
  against embedded Dolt.

Drop TestImportChunkedConcurrentAvailability: it probed SQLite two-connection
file-lock fairness, which does not exist on main (embedded Dolt serializes
writes in-process and refcount-aliases handles per data dir), so it cannot be
ported faithfully. The inter-chunk pause it validated end-to-end stays covered
by TestImportChunkedPausesBetweenChunkTransactions. Removing it also clears the
sole testing.Short() the policy check rejected.
