+++ cmd/bd/import_chunking_test.go
@@ -5,14 +5,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"path/filepath"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/steveyegge/beads/internal/storage"
-	"github.com/steveyegge/beads/internal/storage/sqlite"
 	"github.com/steveyegge/beads/internal/types"
 )
 
@@ -20,6 +17,11 @@ import (
 // snapshot (issue rows plus the dependencies attached at call time), so tests
 // can assert on transaction boundaries. failOnCall simulates a mid-import
 // failure at the Nth call (1-based; 0 = never fail).
+//
+// Tests that need a real, persisting engine (durable committed prefixes,
+// readiness recomputation, event dedup) live in import_chunking_embedded_test.go
+// and run against embedded Dolt behind BEADS_TEST_EMBEDDED_DOLT; the shared
+// helpers below are reused by both files.
 type chunkRecordingStore struct {
 	storage.DoltStorage
 	batches    [][]*types.Issue
@@ -309,180 +311,6 @@ func TestImportChunkedPausesBetweenChunkTransactions(t *testing.T) {
 	}
 }
 
-// failNthCreateStore wraps a real store and fails the Nth batch-create call,
-// simulating a crash mid-import against a real engine.
-type failNthCreateStore struct {
-	storage.DoltStorage
-	calls      int
-	failOnCall int
-}
-
-func (f *failNthCreateStore) CreateIssuesWithFullOptions(ctx context.Context, issues []*types.Issue, actor string, opts storage.BatchCreateOptions) error {
-	f.calls++
-	if f.calls == f.failOnCall {
-		return errors.New("simulated crash")
-	}
-	return f.DoltStorage.CreateIssuesWithFullOptions(ctx, issues, actor, opts)
-}
-
-func provisionChunkStore(t *testing.T) storage.DoltStorage {
-	t.Helper()
-	ctx := context.Background()
-	store, err := sqlite.Provision(ctx, filepath.Join(t.TempDir(), "beads.db"))
-	if err != nil {
-		t.Fatalf("provision sqlite store: %v", err)
-	}
-	t.Cleanup(func() { _ = store.Close() })
-	if err := store.SetConfig(ctx, "issue_prefix", "bd"); err != nil {
-		t.Fatalf("seed issue_prefix: %v", err)
-	}
-	return store
-}
-
-// End-to-end against the real SQLite engine: a mid-import failure leaves a
-// committed, queryable prefix, and re-running the same import converges —
-// every row present, cross-chunk dependencies wired, no duplicated events.
-func TestImportChunkedRealStoreResumeConverges(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-	store := provisionChunkStore(t)
-
-	makeIssues := func() []*types.Issue {
-		issues := chunkTestIssues(12)
-		// Blocking forward reference in file order: 1 -> 11.
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepBlocks}}
-		return issues
-	}
-
-	failing := &failNthCreateStore{DoltStorage: store, failOnCall: 2}
-	_, err := importIssuesCore(ctx, "", failing, makeIssues(), ImportOptions{SkipPrefixValidation: true})
-	if err == nil {
-		t.Fatalf("importIssuesCore succeeded, want simulated crash at chunk 2")
-	}
-
-	// Some rows are committed and queryable, others absent: the transactions
-	// really are bounded.
-	committed, absent := 0, 0
-	for i := 1; i <= 12; i++ {
-		if _, err := store.GetIssue(ctx, fmt.Sprintf("bd-chunk%02d", i)); err == nil {
-			committed++
-		} else {
-			absent++
-		}
-	}
-	if committed != 5 || absent != 7 {
-		t.Fatalf("after crash at chunk 2: committed=%d absent=%d, want exactly the first chunk (5) durable", committed, absent)
-	}
-
-	// Re-run the identical import: it must converge.
-	result, err := importIssuesCore(ctx, "", store, makeIssues(), ImportOptions{SkipPrefixValidation: true})
-	if err != nil {
-		t.Fatalf("re-run importIssuesCore: %v", err)
-	}
-	if result.Created != 12 {
-		t.Fatalf("re-run Created = %d, want all 12 rows accounted for", result.Created)
-	}
-	for i := 1; i <= 12; i++ {
-		id := fmt.Sprintf("bd-chunk%02d", i)
-		if _, err := store.GetIssue(ctx, id); err != nil {
-			t.Fatalf("issue %s missing after re-run: %v", id, err)
-		}
-	}
-	deps, err := store.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies: %v", err)
-	}
-	if len(deps) != 1 || deps[0].ID != "bd-chunk11" {
-		t.Fatalf("cross-chunk dependency not wired, got %#v", deps)
-	}
-	// Rows upserted twice must not accrue duplicate created events.
-	for _, id := range []string{"bd-chunk01", "bd-chunk02"} {
-		events, err := store.GetEvents(ctx, id, 50)
-		if err != nil {
-			t.Fatalf("GetEvents(%s): %v", id, err)
-		}
-		created := 0
-		for _, e := range events {
-			if e.EventType == types.EventCreated {
-				created++
-			}
-		}
-		if created != 1 {
-			t.Fatalf("%s has %d created events after resume, want 1", id, created)
-		}
-	}
-}
-
-// No freeze point of a chunked import may expose a blocked bead as ready:
-// whenever an imported row is visible to a concurrent reader, the blocking
-// edges its import file declares must be visible too. This is the red-team
-// ready-window finding: the previous two-phase layout committed every row
-// dep-less first, so `bd ready` mid-import offered blocked work for dispatch.
-func TestImportChunkedNoReadyWindowAtAnyFreezePoint(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-
-	makeIssues := func() []*types.Issue {
-		issues := chunkTestIssues(12)
-		// bd-chunk01 is BLOCKED by bd-chunk11 (forward ref in file order).
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepBlocks}}
-		return issues
-	}
-
-	// Learn how many transactions a full import issues.
-	full := provisionChunkStore(t)
-	counting := &failNthCreateStore{DoltStorage: full}
-	if _, err := importIssuesCore(ctx, "", counting, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("full import: %v", err)
-	}
-	totalCalls := counting.calls
-	if totalCalls < 2 {
-		t.Fatalf("totalCalls = %d, want a chunked import", totalCalls)
-	}
-	ready, err := full.GetReadyWork(ctx, types.WorkFilter{Limit: 50})
-	if err != nil {
-		t.Fatalf("GetReadyWork: %v", err)
-	}
-	for _, issue := range ready {
-		if issue.ID == "bd-chunk01" {
-			t.Fatalf("bd-chunk01 ready after a full import; it is blocked by open bd-chunk11")
-		}
-	}
-
-	// Freeze the store at every possible transaction boundary and check the
-	// invariant: row visible => blocking edge visible => not ready.
-	for failAt := 1; failAt <= totalCalls; failAt++ {
-		store := provisionChunkStore(t)
-		failing := &failNthCreateStore{DoltStorage: store, failOnCall: failAt}
-		if _, err := importIssuesCore(ctx, "", failing, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err == nil {
-			t.Fatalf("freeze point %d: import unexpectedly succeeded", failAt)
-		}
-		if _, err := store.GetIssue(ctx, "bd-chunk01"); err != nil {
-			continue // row not committed yet: nothing to observe
-		}
-		deps, err := store.GetDependencies(ctx, "bd-chunk01")
-		if err != nil {
-			t.Fatalf("freeze point %d: GetDependencies: %v", failAt, err)
-		}
-		if len(deps) == 0 {
-			t.Fatalf("freeze point %d: bd-chunk01 committed without its blocking edge — ready window", failAt)
-		}
-		ready, err := store.GetReadyWork(ctx, types.WorkFilter{Limit: 50})
-		if err != nil {
-			t.Fatalf("freeze point %d: GetReadyWork: %v", failAt, err)
-		}
-		for _, issue := range ready {
-			if issue.ID == "bd-chunk01" {
-				t.Fatalf("freeze point %d: blocked bead bd-chunk01 offered as ready work mid-import", failAt)
-			}
-		}
-	}
-}
-
 // orderImportIssuesForChunking must emit a cycle member before a valid row that
 // blocks on it, even when that row precedes the cycle in file order. A plain
 // file-order cycle fallback would chunk the dependent ahead of its blocker and
@@ -506,476 +334,5 @@ func TestOrderImportIssuesForChunkingPlacesCycleBeforeDependent(t *testing.T) {
 	if pos["bd-chunk03"] > pos["bd-chunk01"] {
 		t.Fatalf("bd-chunk03 (blocker on a cycle) at %d must precede its dependent bd-chunk01 at %d so the edge rides inline",
 			pos["bd-chunk03"], pos["bd-chunk01"])
-	}
-}
-
-// A readiness edge that points INTO a tolerated readiness cycle must still ride
-// inline with its row across every import freeze point: the cycle-fallback
-// ordering must not place a valid dependent of a cycle in an earlier chunk than
-// the cycle member it blocks on. The import only meets a blocking cycle in
-// corrupted or legacy JSONL, which it tolerates (SkipDependencyValidationErrors),
-// but the no-ready-window invariant must hold there too — otherwise the
-// dependent commits ready-without-blocker for the rest of the import. Regression
-// for the attempt-1 review finding.
-func TestImportChunkedNoReadyWindowWithReadinessEdgeIntoCycle(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-
-	makeIssues := func() []*types.Issue {
-		issues := chunkTestIssues(12)
-		// bd-chunk11 <-> bd-chunk12 form a tolerated blocking cycle. bd-chunk01
-		// validly blocks on bd-chunk11; in file order it precedes the cycle, so
-		// a plain file-order fallback would chunk it ahead of bd-chunk11 and
-		// defer the live edge.
-		issues[10].Dependencies = []*types.Dependency{{IssueID: issues[10].ID, DependsOnID: issues[11].ID, Type: types.DepBlocks}}
-		issues[11].Dependencies = []*types.Dependency{{IssueID: issues[11].ID, DependsOnID: issues[10].ID, Type: types.DepBlocks}}
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepBlocks}}
-		return issues
-	}
-
-	// A full import must tolerate the cycle, leave bd-chunk01 blocked by open
-	// bd-chunk11, and report how many transactions the import issues.
-	full := provisionChunkStore(t)
-	counting := &failNthCreateStore{DoltStorage: full}
-	if _, err := importIssuesCore(ctx, "", counting, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("full import (tolerated cycle): %v", err)
-	}
-	totalCalls := counting.calls
-	if totalCalls < 2 {
-		t.Fatalf("totalCalls = %d, want a chunked import", totalCalls)
-	}
-	deps, err := full.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies(bd-chunk01): %v", err)
-	}
-	if len(deps) != 1 || deps[0].ID != "bd-chunk11" {
-		t.Fatalf("bd-chunk01 must be blocked by bd-chunk11 after import, got %#v", deps)
-	}
-	ready, err := full.GetReadyWork(ctx, types.WorkFilter{Limit: 50})
-	if err != nil {
-		t.Fatalf("GetReadyWork: %v", err)
-	}
-	for _, issue := range ready {
-		if issue.ID == "bd-chunk01" {
-			t.Fatalf("bd-chunk01 ready after a full import; it is blocked by open bd-chunk11 (cycle member)")
-		}
-	}
-
-	// Freeze at every transaction boundary: whenever bd-chunk01 is visible, its
-	// blocking edge into the cycle must be visible too, so it is never offered
-	// as ready mid-import.
-	for failAt := 1; failAt <= totalCalls; failAt++ {
-		store := provisionChunkStore(t)
-		failing := &failNthCreateStore{DoltStorage: store, failOnCall: failAt}
-		if _, err := importIssuesCore(ctx, "", failing, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err == nil {
-			t.Fatalf("freeze point %d: import unexpectedly succeeded", failAt)
-		}
-		if _, err := store.GetIssue(ctx, "bd-chunk01"); err != nil {
-			continue // row not committed yet: nothing to observe
-		}
-		deps, err := store.GetDependencies(ctx, "bd-chunk01")
-		if err != nil {
-			t.Fatalf("freeze point %d: GetDependencies: %v", failAt, err)
-		}
-		if len(deps) == 0 {
-			t.Fatalf("freeze point %d: bd-chunk01 committed without its blocking edge into the cycle — ready window", failAt)
-		}
-		ready, err := store.GetReadyWork(ctx, types.WorkFilter{Limit: 50})
-		if err != nil {
-			t.Fatalf("freeze point %d: GetReadyWork: %v", failAt, err)
-		}
-		for _, issue := range ready {
-			if issue.ID == "bd-chunk01" {
-				t.Fatalf("freeze point %d: blocked bead bd-chunk01 offered as ready work mid-import", failAt)
-			}
-		}
-	}
-}
-
-// A waits-for waiter's is_blocked state is gated on its spawner having an active
-// parent-child child, and the per-chunk is_blocked recompute only re-evaluates
-// the rows in its own transaction. So when the waiter and its spawner's active
-// child straddle a chunk boundary in file order (waiter first, child later), a
-// naive ordering leaves the waiter's chunk computing is_blocked=0 against a
-// still-childless spawner, and the later chunk that imports the child never
-// re-blocks the waiter — it sits ready for the rest of the import and after it.
-// Ordering the spawner inline with the waiter is not enough; every in-batch
-// child of a waited spawner must be emitted no later than the waiter. Regression
-// for the attempt-2 review blocker.
-func TestImportChunkedWaitsForChildImportedLaterStaysBlocked(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-
-	makeIssues := func() []*types.Issue {
-		// File order: waiter, spawner, three fillers, active child. At chunk
-		// size 5 the waiter (chunk 0) and the child (chunk 1) straddle the
-		// boundary unless the import reorders the child ahead of the waiter.
-		// bd-chunk01 waits for bd-chunk02; bd-chunk06 is an open parent-child
-		// child of bd-chunk02.
-		issues := chunkTestIssues(6)
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[1].ID, Type: types.DepWaitsFor}}
-		issues[5].Dependencies = []*types.Dependency{{IssueID: issues[5].ID, DependsOnID: issues[1].ID, Type: types.DepParentChild}}
-		return issues
-	}
-
-	full := provisionChunkStore(t)
-	counting := &failNthCreateStore{DoltStorage: full}
-	if _, err := importIssuesCore(ctx, "", counting, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("full import (waits-for child across chunks): %v", err)
-	}
-	if counting.calls < 2 {
-		t.Fatalf("calls = %d, want a chunked (>=2 transaction) import so the waiter and child can straddle a boundary", counting.calls)
-	}
-
-	// The waits-for edge must be wired inline so the gate can see it.
-	waiterDeps, err := full.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies(bd-chunk01): %v", err)
-	}
-	if len(waiterDeps) != 1 || waiterDeps[0].ID != "bd-chunk02" {
-		t.Fatalf("bd-chunk01 must wait for bd-chunk02 after import, got %#v", waiterDeps)
-	}
-
-	// The waiter must NOT be ready: its spawner bd-chunk02 has an open child
-	// bd-chunk06, so the waits-for gate keeps bd-chunk01 blocked.
-	ready, err := full.GetReadyWork(ctx, types.WorkFilter{Limit: 50})
-	if err != nil {
-		t.Fatalf("GetReadyWork: %v", err)
-	}
-	for _, issue := range ready {
-		if issue.ID == "bd-chunk01" {
-			t.Fatalf("bd-chunk01 ready after a full import; it waits for bd-chunk02 whose active child bd-chunk06 was imported in a later chunk")
-		}
-	}
-}
-
-// raceInjectingStore simulates a concurrent writer landing between import
-// transactions: before the Nth CreateIssuesWithFullOptions call it updates an
-// issue directly, bumping its updated_at — exactly what a concurrent
-// `bd update` or a gc claim does in the windows the chunking opens up.
-type raceInjectingStore struct {
-	storage.DoltStorage
-	calls        int
-	injectBefore int
-	injectID     string
-	injected     bool
-	t            *testing.T
-}
-
-func (r *raceInjectingStore) CreateIssuesWithFullOptions(ctx context.Context, issues []*types.Issue, actor string, opts storage.BatchCreateOptions) error {
-	r.calls++
-	if r.calls == r.injectBefore && !r.injected {
-		r.injected = true
-		if err := r.DoltStorage.UpdateIssue(ctx, r.injectID, map[string]interface{}{"assignee": "concurrent-racer"}, "racer"); err != nil {
-			r.t.Fatalf("inject concurrent update: %v", err)
-		}
-	}
-	return r.DoltStorage.CreateIssuesWithFullOptions(ctx, issues, actor, opts)
-}
-
-// A concurrent update to a row between its commit and the deferred-dependency
-// pass must not drop the import's edges for it. This is the red-team #1
-// finding: the old dependency pass resubmitted the full row with
-// RejectStaleUpserts, so the now-newer stored row stale-rejected the resubmit
-// and the engine dropped its deps — silently, permanently (re-runs pre-filter
-// the row away). The dependency pass must wire edges for every row phase 1
-// accepted, regardless of later row-level updates, and must not clobber the
-// concurrent write.
-func TestImportChunkedConcurrentUpdateKeepsDeferredDeps(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-	store := provisionChunkStore(t)
-
-	makeIssues := func() []*types.Issue {
-		issues := chunkTestIssues(12)
-		// A non-blocking forward edge stays deferred (related edges do not
-		// affect readiness, so the ordering pass leaves file order alone):
-		// this is the shape that still crosses the inter-transaction window.
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepRelated}}
-		return issues
-	}
-
-	// 12 issues at chunk size 5 = 3 row chunks (calls 1-3); the deferred
-	// dependency pass is call 4. Inject the concurrent update just before it.
-	racing := &raceInjectingStore{DoltStorage: store, injectBefore: 4, injectID: "bd-chunk01", t: t}
-	result, err := importIssuesCore(ctx, "", racing, makeIssues(), ImportOptions{SkipPrefixValidation: true})
-	if err != nil {
-		t.Fatalf("importIssuesCore: %v", err)
-	}
-	if !racing.injected {
-		t.Fatalf("race was never injected (calls=%d); test harness broken", racing.calls)
-	}
-
-	// The concurrent write survives: the dependency pass must not rewrite rows.
-	got, err := store.GetIssue(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("bd-chunk01 not committed: %v", err)
-	}
-	if got.Assignee != "concurrent-racer" {
-		t.Fatalf("assignee = %q, want the concurrent update preserved", got.Assignee)
-	}
-
-	// The import's edge survives: phase 1 accepted the row, so its deferred
-	// deps must be wired even though the stored row is now newer.
-	deps, err := store.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies: %v", err)
-	}
-	if len(deps) != 1 || deps[0].ID != "bd-chunk11" {
-		t.Fatalf("deferred edge dropped after concurrent update: deps = %#v", deps)
-	}
-
-	// And the reporting must not misclassify the row: its row IS committed.
-	for _, id := range result.StaleSkippedIDs {
-		if id == "bd-chunk01" {
-			t.Fatalf("bd-chunk01 reported stale-skipped, but its row was imported in phase 1")
-		}
-	}
-	if result.Created != 12 {
-		t.Fatalf("Created = %d, want 12", result.Created)
-	}
-}
-
-// The red-team #1 repro shape verbatim, for blocking edges: a claim-like
-// update racing the import must leave the blocking edge intact. With ordered
-// inline wiring the row and its edge commit atomically, so no interleaving of
-// row-level updates can separate them.
-func TestImportChunkedConcurrentUpdateKeepsBlockingDeps(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-	store := provisionChunkStore(t)
-
-	makeIssues := func() []*types.Issue {
-		issues := chunkTestIssues(12)
-		issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepBlocks}}
-		return issues
-	}
-
-	// The blocker's row (bd-chunk11) lands before bd-chunk01's chunk. Update
-	// it between chunks — the interleaving that used to poison the dep pass.
-	racing := &raceInjectingStore{DoltStorage: store, injectBefore: 3, injectID: "bd-chunk11", t: t}
-	if _, err := importIssuesCore(ctx, "", racing, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("importIssuesCore: %v", err)
-	}
-	if !racing.injected {
-		t.Fatalf("race was never injected (calls=%d); test harness broken", racing.calls)
-	}
-	deps, err := store.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies: %v", err)
-	}
-	if len(deps) != 1 || deps[0].ID != "bd-chunk11" {
-		t.Fatalf("blocking edge missing after concurrent update: deps = %#v", deps)
-	}
-
-	// Re-running the identical import stays convergent and keeps the edge.
-	if _, err := importIssuesCore(ctx, "", store, makeIssues(), ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("re-run importIssuesCore: %v", err)
-	}
-	deps, err = store.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies after re-run: %v", err)
-	}
-	if len(deps) != 1 {
-		t.Fatalf("blocking edge lost on re-run: deps = %#v", deps)
-	}
-}
-
-// The other half of the stale-snapshot invariant (bd-578h9.8): when the row
-// write itself was stale-rejected in phase 1 (a local update landed between
-// the pre-filter read and the row's chunk), the snapshot's deferred deps must
-// stay out too. This is why the dependency pass cannot simply run everything
-// with ConflictSkip: it must be restricted to rows phase 1 accepted.
-func TestImportChunkedStaleRejectedRowKeepsDeferredDepsOut(t *testing.T) {
-	setImportChunkSize(t, 5)
-	recordImportPauses(t)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-	store := provisionChunkStore(t)
-	base := time.Date(2026, 6, 1, 12, 0, 0, 0, time.UTC)
-
-	// Seed bd-chunk01 with an OLDER snapshot so the incoming row passes the
-	// pre-filter (incoming strictly newer than local at read time).
-	seed := &types.Issue{ID: "bd-chunk01", Title: "seeded", UpdatedAt: base.Add(-time.Hour)}
-	seed.SetDefaults()
-	if _, err := importIssuesCore(ctx, "", store, []*types.Issue{seed}, ImportOptions{SkipPrefixValidation: true}); err != nil {
-		t.Fatalf("seed import: %v", err)
-	}
-
-	issues := chunkTestIssues(12)
-	issues[0].Dependencies = []*types.Dependency{{IssueID: issues[0].ID, DependsOnID: issues[10].ID, Type: types.DepRelated}}
-
-	// Bump bd-chunk01 to "now" after the pre-filter read but before its row
-	// chunk: the in-transaction stale guard rejects the row write, so the
-	// whole snapshot — deferred deps included — must stay out.
-	racing := &raceInjectingStore{DoltStorage: store, injectBefore: 1, injectID: "bd-chunk01", t: t}
-	result, err := importIssuesCore(ctx, "", racing, issues, ImportOptions{SkipPrefixValidation: true})
-	if err != nil {
-		t.Fatalf("importIssuesCore: %v", err)
-	}
-	if !racing.injected {
-		t.Fatalf("race was never injected; test harness broken")
-	}
-
-	deps, err := store.GetDependencies(ctx, "bd-chunk01")
-	if err != nil {
-		t.Fatalf("GetDependencies: %v", err)
-	}
-	if len(deps) != 0 {
-		t.Fatalf("stale-rejected row's deferred deps were wired anyway: %#v (violates bd-578h9.8)", deps)
-	}
-	found := false
-	for _, id := range result.StaleSkippedIDs {
-		if id == "bd-chunk01" {
-			found = true
-		}
-	}
-	if !found {
-		t.Fatalf("bd-chunk01 not reported stale-skipped; StaleSkippedIDs = %v", result.StaleSkippedIDs)
-	}
-}
-
-// timestampingStore records when each import transaction starts and returns,
-// so the availability test can reason about "strictly inside the import".
-type timestampingStore struct {
-	storage.DoltStorage
-	mu     sync.Mutex
-	starts []time.Time
-	ends   []time.Time
-}
-
-func (s *timestampingStore) CreateIssuesWithFullOptions(ctx context.Context, issues []*types.Issue, actor string, opts storage.BatchCreateOptions) error {
-	s.mu.Lock()
-	s.starts = append(s.starts, time.Now())
-	s.mu.Unlock()
-	err := s.DoltStorage.CreateIssuesWithFullOptions(ctx, issues, actor, opts)
-	s.mu.Lock()
-	s.ends = append(s.ends, time.Now())
-	s.mu.Unlock()
-	return err
-}
-
-// The point of chunking is availability: concurrent bd operations must keep
-// succeeding WHILE a large import runs, not merely after it. This probes a
-// live chunked import from a second store handle (a separate connection pool
-// on the same database file — the same file-lock domain a separate bd process
-// contends on) and requires reads and writes to complete strictly inside the
-// import window. The red-team measurement showed bounded transactions alone
-// do not deliver this: back-to-back chunks starve waiters; the inter-chunk
-// pause is what makes this test pass.
-func TestImportChunkedConcurrentAvailability(t *testing.T) {
-	if testing.Short() {
-		t.Skip("multi-second availability probe; skipped in -short")
-	}
-	setImportChunkSize(t, 100)
-	setImportProgressBuffer(t)
-	ctx := context.Background()
-	dbPath := filepath.Join(t.TempDir(), "beads.db")
-	importSide, err := sqlite.Provision(ctx, dbPath)
-	if err != nil {
-		t.Fatalf("provision import store: %v", err)
-	}
-	defer importSide.Close()
-	if err := importSide.SetConfig(ctx, "issue_prefix", "bd"); err != nil {
-		t.Fatalf("seed issue_prefix: %v", err)
-	}
-	probeSide, err := sqlite.Provision(ctx, dbPath)
-	if err != nil {
-		t.Fatalf("provision probe store: %v", err)
-	}
-	defer probeSide.Close()
-
-	stop := make(chan struct{})
-	var probeMu sync.Mutex
-	var readOK, writeOK []time.Time
-	var probeErrs []error
-	var wg sync.WaitGroup
-	wg.Add(2)
-	go func() { // reader probe
-		defer wg.Done()
-		for {
-			select {
-			case <-stop:
-				return
-			default:
-			}
-			if _, err := probeSide.GetReadyWork(ctx, types.WorkFilter{Limit: 5}); err != nil {
-				probeMu.Lock()
-				probeErrs = append(probeErrs, fmt.Errorf("ready probe: %w", err))
-				probeMu.Unlock()
-			} else {
-				probeMu.Lock()
-				readOK = append(readOK, time.Now())
-				probeMu.Unlock()
-			}
-			time.Sleep(25 * time.Millisecond)
-		}
-	}()
-	go func() { // writer probe
-		defer wg.Done()
-		for n := 0; ; n++ {
-			select {
-			case <-stop:
-				return
-			default:
-			}
-			issue := &types.Issue{ID: fmt.Sprintf("bd-p%04d", n), Title: "probe"}
-			issue.SetDefaults()
-			if err := probeSide.CreateIssue(ctx, issue, "probe"); err != nil {
-				probeMu.Lock()
-				probeErrs = append(probeErrs, fmt.Errorf("create probe %d: %w", n, err))
-				probeMu.Unlock()
-			} else {
-				probeMu.Lock()
-				writeOK = append(writeOK, time.Now())
-				probeMu.Unlock()
-			}
-			time.Sleep(50 * time.Millisecond)
-		}
-	}()
-
-	ts := &timestampingStore{DoltStorage: importSide}
-	_, err = importIssuesCore(ctx, "", ts, chunkTestIssues(1000), ImportOptions{SkipPrefixValidation: true})
-	close(stop)
-	wg.Wait()
-	if err != nil {
-		t.Fatalf("importIssuesCore: %v", err)
-	}
-	if len(ts.starts) < 4 {
-		t.Fatalf("import issued %d transactions, want a genuinely chunked run", len(ts.starts))
-	}
-
-	for _, err := range probeErrs {
-		t.Errorf("concurrent bd operation failed during import: %v", err)
-	}
-	// "Strictly inside": after the first transaction committed and before the
-	// last one began — the region the old single-transaction import blacked out.
-	windowStart := ts.ends[0]
-	windowEnd := ts.starts[len(ts.starts)-1]
-	if !windowEnd.After(windowStart) {
-		t.Fatalf("degenerate import window [%v, %v]", windowStart, windowEnd)
-	}
-	inWindow := func(times []time.Time) int {
-		n := 0
-		for _, at := range times {
-			if at.After(windowStart) && at.Before(windowEnd) {
-				n++
-			}
-		}
-		return n
-	}
-	if got := inWindow(readOK); got == 0 {
-		t.Errorf("no read completed strictly inside the import window (%d total reads)", len(readOK))
-	}
-	if got := inWindow(writeOK); got == 0 {
-		t.Errorf("no write completed strictly inside the import window (%d total writes)", len(writeOK))
 	}
 }
